requirements.txt
pathway==0.11.0
# openai>=1.0.0
groq>=0.4.0
streamlit>=1.28.0
alpha_vantage
pandas
python-dotenv
watchdog
requests
pypdf
